Update on "using float type to do the computation of norm reduce for cpu half and bfloat16 dtype"
As the title, we should use a higher dtype to compute norm reduce for half and bfloat1 dtype.

cc jgong5 mingfeima sanchitintel ashokei jingxu10

[ghstack-poisoned]
