+++ aten/src/ATen/native/cuda/ReduceNormKernel.cu
@@ -16,25 +16,17 @@ namespace at::native {
 template <typename scalar_t, typename acc_t=typename scalar_value_type<scalar_t>::type, typename out_t=typename scalar_value_type<scalar_t>::type>
 void norm_kernel_cuda_impl(TensorIterator& iter, double p) {
   if (p == static_cast<double>(0)) {
-    gpu_reduce_kernel<scalar_t, out_t>(
-        iter, NormZeroOps<scalar_t, acc_t, out_t>(), 0);
+    gpu_reduce_kernel<scalar_t, out_t>(iter, NormZeroOps<scalar_t, acc_t, out_t>(), 0);
   } else if (p == static_cast<double>(1)) {
-    gpu_reduce_kernel<scalar_t, out_t>(
-        iter, NormOneOps<scalar_t, acc_t, out_t>(), 0);
+    gpu_reduce_kernel<scalar_t, out_t>(iter, NormOneOps<scalar_t, acc_t, out_t>(), 0);
   } else if (p == static_cast<double>(2)) {
-    gpu_reduce_kernel<scalar_t, out_t>(
-        iter, NormTwoOps<scalar_t, acc_t, out_t>(), 0);
+    gpu_reduce_kernel<scalar_t, out_t>(iter, NormTwoOps<scalar_t, acc_t, out_t>(), 0);
   } else if (p == static_cast<double>(INFINITY)) {
-    gpu_reduce_kernel<scalar_t, out_t>(
-        iter, AbsMaxOps<scalar_t, acc_t, out_t>(), 0);
+    gpu_reduce_kernel<scalar_t, out_t>(iter, AbsMaxOps<scalar_t, acc_t, out_t>(), 0);
   } else if (p == static_cast<double>(-INFINITY)) {
-    gpu_reduce_kernel<scalar_t, out_t>(
-        iter,
-        AbsMinOps<scalar_t, acc_t, out_t>(),
-        std::numeric_limits<acc_t>::infinity());
+    gpu_reduce_kernel<scalar_t, out_t>(iter, AbsMinOps<scalar_t, acc_t, out_t>(), std::numeric_limits<acc_t>::infinity());
   } else {
-    gpu_reduce_kernel<scalar_t, out_t>(
-        iter, NormOps<scalar_t, acc_t, out_t>{acc_t(p)}, 0);
+    gpu_reduce_kernel<scalar_t, out_t>(iter, NormOps<scalar_t, acc_t, out_t>{acc_t(p)}, 0);
   }
 }
